Add new Follow action using TDD.
Started by adding the test, which fails indicating a "follow" action is
not found in the User controller.

After adding the code to the controller and the new route to `/follow`,
the test started passing.

$ bin/rails test:controllers

Also added a Follow button to the view, to implement the action in
interactive use.

There are still shortcomings, since it still offers a Follow button even
if you're already following the user, or even if you're the user itself.
It's currently not possible to unfollow an user either.